cleanup dead code, move formatting route code to service
